+++ src/snowflake/snowpark/session.py
@@ -99,7 +99,6 @@ from snowflake.snowpark.functions import (
     to_date,
     to_decimal,
     to_geography,
-    to_geometry,
     to_object,
     to_time,
     to_timestamp,
@@ -118,7 +117,6 @@ from snowflake.snowpark.types import (
     DateType,
     DecimalType,
     GeographyType,
-    GeometryType,
     MapType,
     StringType,
     StructType,
@@ -1575,7 +1573,6 @@ class Session:
                         DateType,
                         TimestampType,
                         GeographyType,
-                        GeometryType,
                     ),
                 )
                 else field.datatype
@@ -1583,7 +1580,7 @@ class Session:
             attrs.append(Attribute(quoted_name, sf_type, field.nullable))
             data_types.append(field.datatype)
 
-        # convert all variant/time/geospatial/array/map data to string
+        # convert all variant/time/geography/array/map data to string
         converted = []
         for row in rows:
             converted_row = []
@@ -1618,8 +1615,6 @@ class Session:
                     converted_row.append(json.dumps(value, cls=PythonObjJSONEncoder))
                 elif isinstance(data_type, GeographyType):
                     converted_row.append(value)
-                elif isinstance(data_type, GeometryType):
-                    converted_row.append(value)
                 else:
                     raise TypeError(
                         f"Cannot cast {type(value)}({value}) to {str(data_type)}."
@@ -1647,8 +1642,6 @@ class Session:
                 project_columns.append(to_variant(parse_json(column(name))).as_(name))
             elif isinstance(field.datatype, GeographyType):
                 project_columns.append(to_geography(column(name)).as_(name))
-            elif isinstance(field.datatype, GeometryType):
-                project_columns.append(to_geometry(column(name)).as_(name))
             elif isinstance(field.datatype, ArrayType):
                 project_columns.append(to_array(parse_json(column(name))).as_(name))
             elif isinstance(field.datatype, MapType):