+++ CHANGELOG.md
@@ -1,12 +1,15 @@
 # Release History
 
 ## 1.6.0 (TBD)
-
 ### New Features
-
 - Added support for aliases for commonly used options in DataFrame reader/writer.
 - Added support for `format(...).load()` and `format(...).save(...)`patterns.
 - Added support for `df.write.csv("path/to/stage")`,`df.write.json("path/to/stage")`,`df.write.parquet("path/to/stage")` patterns to allow a similar experience to dataframe readers.
+- Added support for Geometry data.
+
+### Dependency updates
+- Updated `snowflake-connector-python` to 3.0.4.
+
 
 ## 1.5.1 (2023-06-19)